chore: store working files before reboot
