+++ client/package.json
@@ -14,6 +14,7 @@
     "@headlessui/react": "^1.7.17",
     "class-variance-authority": "^0.7.0",
     "clsx": "^2.0.0",
+    "immer": "^10.0.3",
     "next": "latest",
     "next-themes": "^0.2.1",
     "react": "latest",
