+++ client/src/app/layout.tsx
@@ -2,7 +2,7 @@ import './globals.css'
 import type { Metadata } from 'next'
 import { Plus_Jakarta_Sans } from 'next/font/google'
 import Navbar from '@/components/Navbar'
-import { ThemeProvider } from '@components/Theme'
+import { ThemeProvider } from '@/components/Theme'
 
 const jakarta = Plus_Jakarta_Sans({ subsets: ['latin'] })
